Real Time Calculator using HTML, CSS, and JS
we are going to discuss how to develop a Real-Time Calculator using HTML, CSS, and JavaScript. Usually, if we observe any real-time calculator we know 

A grid of numbers (0-9 and 00).
Some basic athematic operators (+, -, /, x, %).
And some symbols for special operations such as (clear, backspace, equal)
The output of our desired program will be as shown below.